+++ CHANGELOG.md
@@ -58,5 +58,9 @@ Added a bunch more error diagnostics. If there's a particular error that catches
 
 ### [2.0.7]
 
-Refactor diagnostic parser to improve maintainability and resolve some false positive errors
-Correct ranges for some commands
+Refactor diagnostic parser to improve performance, maintainability and resolve some false positive errors
+Improve parser to understand multiline string parameters
+Add error for when an event command requires a site name to be specified in the #msg
+Extend validation to cover all modding commands
+Using a nonexistent command is an error
+Using a command outside of its intended scope is an error